adding limb-based generalized posits (#310)
* bumping SEMVER to v3.61

* WIP: reimplementing posits with blocks and adding generalization

* WIP: adding alignment spec to posit2

* adding simple API test to bring up posit2

* code hygiene for blocksignificant

* add missing file

* adding missing numeric_limits.hpp

* adding all/any/none/count API to blockbinary

* adding new BUILD_CI target to bring back the compilation time for Github Actions

* adding missing manipulators.hpp

* internal API change to get raw bits

* WIP: checkpointing code mods

* WIP: adding specific tests for attribute and manipulator functions to create better coverage for MSVC vs GCC vs CLang compilation differences

* WIP: enabling constexpr on generalized posits

* quick fix for LONG_DOUBLE_SUPPORT for GCC and CLang

* code hygiene on posit, adaptiveposit, blocksignificant, cfloat, and value

* fixing gcc array_bounds warnings

* parameterizing conversion functions for generalized posits

* trying to set GCC compiler version in CI

* trying to get GCC compiler to move off 9 in CI

* dropping to Ubuntu 20.04 to see if CI gets unstuck

* updating devcontainer from Ubuntu 20 to 22, and gcc-10 to gcc-11

* code hygiene to try to isolate the gcc array-bound warnings

* refining the regression set

* fix to workflow yaml

* fix to workflow yaml

* fix to workflow yaml

* setup-cpp-compiler script.sh added

* constexpr trivial construction fixes

* reducing the size of the regression matrix

* adding regression status badge

* fixing badge URL

* adding development branch status badge

* rearranging badges and leading paragraph in README.md

* WIP: completing attributes test for generalized posits

* restructuring build and organization to unifiy number and arithmetic traits

* removing superfluous attribute specializations now that we have a common format for number and arithmetic traits

* WIP: unifying number and arithmetic traits across all number systems

* change API to common arithmetic trades to support native types

* removing clang-9 from Unbuntu, adding clang-11 to MacOS os/compiler matrix

* enabling regression test for lns arithmetic exceptions

* adding a number and type traits test for generalized posits

* compilation fix for lns multiplication test

* removing MacOS clang-11 from CI as it falls back to clang-13

* adding traits test to classic floating-point

* bug fix of operator-() when lns value is special encoding of 0

* adding ulp calculus for generalized posits

* refining the ULP math api demonstration for regular and generalized posits

* adding ulp demo to lns

* making generalized posits trivially constructable

* WIP: debugging float->posit conversion

* refining BlockTripleOperator before using it across all generalized posits

* compilation fix for GCC and Clang

* changing the LNS template interface to use unsigned int instead of size_t for non-type parameters

* changing the FIXPNT template interface to use unsigned int instead of size_t for non-type parameters

* changing the INTEGER template interface to use unsigned int instead of size_t for non-type parameters

* WIP: changing the template interface to use unsigned int instead of size_t for non-type parameters

* changing the CFLOAT template interface to use unsigned int instead of size_t for non-type parameters

* changing AREAL and POSIT template contract to use unsigned instead of size_t for non-type parameters

* compilation fixes for non-type template parameter type reassignment

* bug fix in setbits() in areal caused by retyping to unsigned

* WIP: compilation fixes for gcc

* WIP: more compilation fixes for gcc

* adding generic randoms and making REGRESSION_LEVEL_STRESS compilable

* code hygiene

* compilation fixes for gcc 8

* quick edit of main branch merge

* code hygiene for clang

+++ applications/blas/l3_fused_mm.cpp
@@ -18,7 +18,7 @@ template<typename Scalar>
 std::string conditional_fdp(const sw::universal::blas::vector< Scalar >& a, const sw::universal::blas::vector< Scalar >& b) {
 	return std::string("no FDP for non-posit value_type");
 }
-template<size_t nbits, size_t es>
+template<unsigned nbits, unsigned es>
 std::string conditional_fdp(const sw::universal::blas::vector< sw::universal::posit<nbits, es> >& a, const sw::universal::blas::vector< sw::universal::posit<nbits, es> >& b) {
 	std::stringstream ss;
 	ss << sw::universal::fdp(a, b);
