+++ include/universal/number/fixpnt/manipulators.hpp
@@ -1,7 +1,7 @@
 #pragma once
 // manipulators.hpp: definition of manipulation functions for fixed-point types
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <sstream>
@@ -9,7 +9,7 @@
 namespace sw { namespace universal {
 
 // Generate a type tag for general fixpnt
-template<size_t nbits, size_t rbits, bool arithmetic, typename bt>
+template<unsigned nbits, unsigned rbits, bool arithmetic, typename bt>
 std::string type_tag(const fixpnt<nbits, rbits, arithmetic, bt>& = {}) {
 	std::stringstream s;
 	s << "fixpnt<"
