+++ include/universal/number/lns/math/fractional.hpp
@@ -7,7 +7,7 @@
 
 namespace sw { namespace universal {
 
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> lnsmod(lns<nbits, rbits, bt, xtra...> x, lns<nbits, rbits, bt, xtra...> y) {
 	using Real = lns<nbits, rbits, bt, xtra...>;
 	if (y.iszero() || x.isinf() || x.isnan() || y.isnan()) {
@@ -34,13 +34,13 @@ lns<nbits, rbits, bt, xtra...> lnsmod(lns<nbits, rbits, bt, xtra...> x, lns<nbit
 }
 
 // fmod retuns x - n*y where n = x/y with the fractional part truncated
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> fmod(lns<nbits, rbits, bt, xtra...> x, lns<nbits, rbits, bt, xtra...> y) {
 	return lnsmod(x, y);
 }
 
 // shim to stdlib
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> remainder(lns<nbits, rbits, bt, xtra...> x, lns<nbits, rbits, bt, xtra...> y) {
 	return lns<nbits, rbits, bt, xtra...>(std::remainder(double(x), double(y)));
 }
@@ -48,7 +48,7 @@ lns<nbits, rbits, bt, xtra...> remainder(lns<nbits, rbits, bt, xtra...> x, lns<n
 // TODO: validate the rounding of these conversion, versus a method that manipulates the fraction bits directly
 
 // frac returns the fraction of a lns value that is > 1
-template<size_t nbits, size_t rbits, typename bt, auto... xtra>
+template<unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 lns<nbits, rbits, bt, xtra...> frac(lns<nbits, rbits, bt, xtra...> x) {
 	using Real = lns<nbits, rbits, bt, xtra...>;
 	long long intValue = (long long)(x);
