+++ include/universal/number/lns/numeric_limits.hpp
@@ -7,7 +7,7 @@
 
 namespace std {
 
-template <size_t nbits, size_t rbits, typename bt, auto... xtra>
+template <unsigned nbits, unsigned rbits, typename bt, auto... xtra>
 class numeric_limits< sw::universal::lns<nbits, rbits, bt, xtra...> > {
 public:
 	using LNS = sw::universal::lns<nbits, rbits, bt, xtra...>;
@@ -43,7 +43,7 @@ public:
 	}
 
 	static constexpr int digits       = -LNS::min_exponent + rbits;
-	static constexpr int digits10     = digits / 3.3;
+	static constexpr int digits10     = static_cast<int>(digits / 3.3f);
 	static constexpr int max_digits10 = digits10;
 	static constexpr bool is_signed   = true;
 	static constexpr bool is_integer  = false;
@@ -51,9 +51,9 @@ public:
 	static constexpr int radix        = 2;
 
 	static constexpr int min_exponent = LNS::min_exponent;
-	static constexpr int min_exponent10 = min_exponent / 3.3;
+	static constexpr int min_exponent10 = static_cast<int>(min_exponent / 3.3f);
 	static constexpr int max_exponent = LNS::max_exponent;
-	static constexpr int max_exponent10 = max_exponent / 3.3;
+	static constexpr int max_exponent10 = static_cast<int>(max_exponent / 3.3f);
 	static constexpr bool has_infinity = false;
 	static constexpr bool has_quiet_NaN = false;
 	static constexpr bool has_signaling_NaN = false;
