+++ include/universal/number/posit/fdp.hpp
@@ -1,7 +1,7 @@
 #pragma once
 // fdp.hpp :  include file containing templated C++ interfaces to fused dot product
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <iostream>
@@ -29,9 +29,9 @@ void fdp_qc(Qy& sum_of_products, size_t n, const Vector& x, size_t incx, const V
 template<typename Vector>
 enable_if_posit<value_type<Vector>, value_type<Vector> > // as return type
 fdp_stride(size_t n, const Vector& x, size_t incx, const Vector& y, size_t incy) {
-	constexpr size_t nbits = Vector::value_type::nbits;
-	constexpr size_t es = Vector::value_type::es;
-	constexpr size_t capacity = 20; // support vectors up to 1M elements
+	constexpr unsigned nbits = Vector::value_type::nbits;
+	constexpr unsigned es = Vector::value_type::es;
+	constexpr unsigned capacity = 20; // support vectors up to 1M elements
 	quire<nbits, es, capacity> q = 0;
 	size_t ix, iy;
 	for (ix = 0, iy = 0; ix < n && iy < n; ix = ix + incx, iy = iy + incy) {
@@ -49,9 +49,9 @@ fdp_stride(size_t n, const Vector& x, size_t incx, const Vector& y, size_t incy)
 template<typename Vector>
 enable_if_posit<value_type<Vector>, value_type<Vector> > // as return type
 fdp(const Vector& x, const Vector& y) {
-	constexpr size_t nbits = Vector::value_type::nbits;
-	constexpr size_t es = Vector::value_type::es;
-	constexpr size_t capacity = 20; // support vectors up to 1M elements
+	constexpr unsigned nbits = Vector::value_type::nbits;
+	constexpr unsigned es = Vector::value_type::es;
+	constexpr unsigned capacity = 20; // support vectors up to 1M elements
 	quire<nbits, es, capacity> q(0);
 	size_t ix, iy, n = size(x);
 	for (ix = 0, iy = 0; ix < n && iy < n; ++ix, ++iy) {
@@ -72,9 +72,9 @@ constexpr auto size(const std::vector<Scalar>& v) -> decltype(v.size())
 template<typename Vector>
 enable_if_posit<value_type<Vector>, value_type<Vector> > // as return type
 fdp(const Vector& x, const Vector& y) {
-	constexpr size_t nbits = Vector::value_type::nbits;
-	constexpr size_t es = Vector::value_type::es;
-	constexpr size_t capacity = 20; // support vectors up to 1M elements
+	constexpr unsigned nbits = Vector::value_type::nbits;
+	constexpr unsigned es = Vector::value_type::es;
+	constexpr unsigned capacity = 20; // support vectors up to 1M elements
 	quire<nbits, es, capacity> q(0);
 	size_t ix, iy, n = size(x);
 	for (ix = 0, iy = 0; ix < n && iy < n; ++ix, ++iy) {
